Refactoring index.js file to improve code maintainability:
  * Create file structure to refactor index.js
  * Move code from index js into additional files for modularity
  * Add additional folder for ai providers
  * Add back helpful comments and misc features
