+++ package.json
@@ -11,7 +11,7 @@
 	"name": "refactorcode",
 	"version": "1.0.0",
 	"description": "A console application to reduce bugs, improve performance and improve readability of your code.",
-	"main": "index.js",
+	"main": "./src/main.js",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1"
 	},
@@ -19,7 +19,7 @@
 	"author": "",
 	"license": "ISC",
 	"bin": {
-		"refactorcode": "./index.js"
+		"refactorcode": "./src/main.js"
 	},
 	"type": "module"
 }
